Parse minimal program
Start with a barebones implementation of parsing an empty program. All nodes are associated with some kind of literal and an offset for error reporting.

The lexer will raise exceptions for invalid tokens that will bubble up. It may be worth having Lexer errors subclass from a general Lexer::Error object that the parser can catch and wrap in its own error.

The lexer, and therefore the parser, will eventually need to accept a fileset, rather than a file, for full multi-file programs.