Parse binary and unary expressions
This involved a significant overhaul to stepping through tokens in the parser to make "lookahead" possible without peeking or backtracking. Infix binary expressions are tricky in that until the operator token is scanned, its not possible to know if its a simple expression or not.

There were other fixes implemented, as well:
 - Exclamation was corrected.
 - Token can report if it is an operator.
 - Node is given defaults.
 - Variable declaration walks its assignment expression.