Create Scope class
Scopes will be needed for variable and type lookup. The class provides
an initializer to set the parent scope, a method to insert, and a method
to lookup types. When lookup returns nil, the variable is not in scope.
A scope with a nil parent is the top-level (global) scope.