Error reporting should use position
This was a huge miss. The offset is file relative. When multi-file support is finished, offset is useless. Instead, the file position object gives the necessary information to locate the actual error.

Additionally, the literal and offset were never used in the error message. This made these values dead code and errors were quite useless. Position is now used in the error message. If the literal value matters, it can be added to the error message.