Switch from offset to position
The offset is pointless in a multi-file situation and using the position makes a lot more sense. This required a ton of changes and improvements.

A bug in the lexer was uncovered as a result, where it was not advancing the offsets in the right way, leading to off by one errors when creating positions.

A massive change to the abstract syntax tree was necessary that made for a much easier to work with interface.

Tests got cleaned up to test what is actually important.